refactor(types): split timer.ts into entities/{timer-config,check-in-record}.ts
